s2n_connection: Move all bitfield fields up front
Ultimately, this is for the benefit of the SAW proofs. Due to a limitation
in how SAW currently works, bitfields must be accessed by index rather than
by name, and due to how often new fields are added to `s2n_connection`, the
only way to do this in way that's maintainable is to have all the bitfield
fields be up front. That way, the index to access the bitfield will always be
zero, which significantly decreases the likelihood that the SAW proofs will
need to be updated with each new field added to `s2n_connection`.

This is all rather unfortunate. See
https://github.com/GaloisInc/saw-script/issues/1461 for a plan to
make handling bitfields more maintainable in SAW.